changed ordinal data type

+++ src/models/catalog_models.rs
@@ -45,7 +45,7 @@ pub struct CatalogCategory {
     pub category_name: String,
     pub parent_category_id: Option<String>,
     pub root_category_id: Option<String>,
-    pub ordinal: i32
+    pub ordinal: i64
 }
 
 #[derive(Serialize, Deserialize, PartialEq, Hash)]
